# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PDAM Tiara Material Tracking & Verification System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branch details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"

  # Materials
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Material details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    put:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"

  # Work Orders
...
  /material-requests:
    get:
      operationId: listMaterialRequests
      tags: [materialRequests]
      summary: List material requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
        - name: requestedBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of material requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialRequest"
    post:
      operationId: createMaterialRequest
      tags: [materialRequests]
      summary: Create a material request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialRequestInput"
      responses:
        "201":
          description: Material request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"
  /material-requests/{id}:
    get:
      operationId: getMaterialRequest
      tags: [materialRequests]
      summary: Get material request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Material request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"
  /material-requests/{id}/approve:
    post:
      operationId: approveMaterialRequest
      tags: [materialRequests]
      summary: Approve a material request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Request approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"
  /material-requests/{id}/reject:
    post:
      operationId: rejectMaterialRequest
      tags: [materialRequests]
      summary: Reject a material request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"

  # Trackings
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # User
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin_gudang, petugas_lapangan, spi, direksi, superadmin]
        branchId:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Branch:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        createdAt:
          type: string
          format: date-time
        activeTrackings:
          type: ["integer", "null"]
        criticalCount:
          type: ["integer", "null"]
    BranchInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string

    # Material
    Material:
      type: object
      required: [id, code, name, category, unit, currentStock, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [valve, pipe, fitting, meter, other]
        unit:
          type: string
        unitPrice:
          type: ["number", "null"]
        currentStock:
          type: integer
        branchId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MaterialInput:
      type: object
      required: [code, name, category, unit]
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [valve, pipe, fitting, meter, other]
        unit:
          type: string
        unitPrice:
          type: ["number", "null"]
        currentStock:
          type: integer
        branchId:
          type: ["string", "null"]

    # Work Order
...
    MaterialRequest:
      type: object
      required: [id, requestNumber, materialId, qtyRequested, status, createdAt]
      properties:
        id:
          type: string
        requestNumber:
          type: string
        workOrderId:
          type: ["string", "null"]
        workOrderNumber:
          type: ["string", "null"]
        materialId:
          type: string
        materialName:
          type: ["string", "null"]
        materialCode:
          type: ["string", "null"]
        qtyRequested:
          type: integer
        requestedById:
          type: ["string", "null"]
        requestedByName:
          type: ["string", "null"]
        approvedById:
          type: ["string", "null"]
        approvedByName:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MaterialRequestInput:
      type: object
      required: [materialId, qtyRequested]
      properties:
        workOrderId:
          type: ["string", "null"]
        materialId:
          type: string
        qtyRequested:
          type: integer
    ApprovalInput:
      type: object
      properties:
        notes:
          type: ["string", "null"]

    # Tracking
...